Generated csv file
Started to address #3 but I still can't match the inputs.
@Vyrwn I spent good amount of time to try to match the values based on
Grasshopper file but I couldn't really figure it out. You can see list
of the inputs and values in scripts/generate_csv.py
If you can come up with a clean list of the values, 4 for each, I can
clean the csv and move forward to the interface.